moved architecture files to subfolder archs to avoid circular imports
